+++ src/main/java/com/coders/explosion/ExplosionGlassMod.java
@@ -18,9 +18,9 @@ import java.io.File;
         guiFactory = "com.coders.explosion.ConfigGuiFactory"
 )
 public class ExplosionGlassMod {
-    public static final String MODID = "explosionglass";
-    public static final String NAME = "ExplosionGlass";
-    public static final String VERSION = "1.9.5";
+    public static final String MODID = "expglass";
+    public static final String NAME = "EXPGlass";
+        public static final String VERSION = "1.9.6";
 
     public static Configuration config;